Moved `WriteLog` to logging module * Additional typo fix for the InputStream.Addon settings
